requirements.txt
XInput-Python>=0.4.0
